code restructure: rename and relocate store related files
