--- src/app/hooks.ts
@@ -1,7 +1,0 @@
-import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
-
-import type { AppDispatch, RootState } from "./store";
-
-// Use throughout your app instead of plain `useDispatch` and `useSelector`
-export const useAppDispatch: () => AppDispatch = useDispatch;
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
