+++ src/components/rawDisplay/RawDisplay.tsx
@@ -7,11 +7,10 @@ import {
 import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 
-import { useAppSelector } from "../../app/hooks";
-import { store } from "../../app/store";
 import { settingsRedux, stopBookmarksRedux } from "../../data/etaObjects";
 import { EtaPredictionXml, RouteXml, RoutesXml } from "../../data/etaXml";
 import { settingsSelectors } from "../../features/settings/settingsSlice";
+import { store, useAppSelector } from "../../store";
 import { fluentStyles } from "../../styles/fluent";
 
 export default function RawDisplay(props: {
