+++ src/features/bookmarks/Bookmark.tsx
@@ -2,11 +2,11 @@ import { Button, Text } from "@fluentui/react-components";
 import { useCallback, useEffect, useState } from "react";
 import { Trans, useTranslation } from "react-i18next";
 
-import { useAppDispatch, useAppSelector } from "../../app/hooks";
 import { FetchXMLWithCancelToken } from "../../components/fetch/fetchUtils";
 import { multiStopParser } from "../../components/parser/multiStopParser";
 import RawDisplay from "../../components/rawDisplay/RawDisplay";
 import { LineStopEta, stopBookmarksRedux } from "../../data/etaObjects";
+import { useAppDispatch, useAppSelector } from "../../store";
 import { BookmarkCard } from "./BookmarkCard";
 import { clearStopBookmarks } from "./stopBookmarkSlice";
