+++ src/features/bookmarks/BookmarkCard.tsx
@@ -5,8 +5,8 @@ import { t } from "i18next";
 import { useCallback } from "react";
 import { Link } from "react-router-dom";
 
-import { useAppDispatch, useAppSelector } from "../../app/hooks";
 import { stopBookmarksRedux } from "../../data/etaObjects";
+import { useAppDispatch, useAppSelector } from "../../store";
 import { fluentStyles } from "../../styles/fluent";
 import { removeStopBookmark } from "./stopBookmarkSlice";
