+++ src/features/eta/BookmarkCardEta.tsx
@@ -2,9 +2,9 @@ import { Badge, Text } from "@fluentui/react-components";
 import { Card, CardHeader } from "@fluentui/react-components/unstable";
 import { Link } from "react-router-dom";
 
-import { useAppSelector } from "../../app/hooks";
 import { CountdownSec } from "../../components/countdown/CountdownSec";
 import { LineStopEta, stopBookmarksRedux } from "../../data/etaObjects";
+import { useAppSelector } from "../../store";
 import { fluentStyles } from "../../styles/fluent";
 
 export function BookmarkCardEta(props: { item: LineStopEta }) {
