+++ src/features/settings/Settings.tsx
@@ -10,10 +10,9 @@ import { FormEvent, useCallback, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { Link } from "react-router-dom";
 
-import { useAppDispatch, useAppSelector } from "../../app/hooks";
-import { store } from "../../app/store";
 import RawDisplay from "../../components/rawDisplay/RawDisplay";
 import { settingsRedux } from "../../data/etaObjects";
+import { store, useAppDispatch, useAppSelector } from "../../store";
 import { fluentStyles } from "../../styles/fluent";
 import { changeSettings, settingsSelectors } from "./settingsSlice";
