+++ src/index.tsx
@@ -3,7 +3,6 @@ import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 import { RouterProvider } from "react-router";
 
-import { store } from "./app/store";
 import { saveStopBookmarkState } from "./features/bookmarks/localstorage";
 import { saveSettingsState } from "./features/settings/localstorage";
 import "./i18n/i18n";
@@ -11,6 +10,7 @@ import "./index.css";
 import reportWebVitals from "./reportWebVitals";
 import { router } from "./routes/MainRouter";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
+import { store } from "./store";
 import { FluentTheme } from "./styles/FluentTheme";
 
 const container = document.getElementById("root");
